pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "findmydevice"
version = "0.10.0"
description = "A Python package for interacting with the findmydevice API."
authors = [{ name = "Dennis Schwertel", email = "s@digitalkultur.net" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
dependencies = [ "argon2-cffi", "attrs", "requests", "cryptography"]

[tool.setuptools]
packages = ["findmydevice"]

[project.optional-dependencies]
dev = ["pytest", "ruff"]


[tool.ruff]
# Enable Ruff linting and formatting
line-length = 88  # Matches Black's default

[tool.ruff.lint]
extend-select = ["I"]  # Enable autofix for import sorting
# Enable recommended linting rules
select = [
    "E",  # Pycodestyle errors
    "W",  # Pycodestyle warnings
    "F",  # Pyflakes
    "I",  # Import sorting
    "B",  # Bugbear (common issues)
    "Q",  # Flake8-quotes (quote consistency)
    "F401", # Unused imports
]

ignore = [
    "E501"  # Ignore line length since Black handles it
]

[tool.ruff.format]
# Enable Ruff formatter
quote-style = "double"
indent-style = "space"
line-ending = "lf"

